backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-genai==1.5.0
tavily-python==0.5.0
upstash-redis==1.1.0
sse-starlette==2.1.0
python-dotenv==1.0.0
